+++ cli/src/pcluster/cli/commands/cluster_logs.py
@@ -74,6 +74,9 @@ class ExportClusterLogsCommand(ExportLogsCommand, CliCommand):
             is_pcluster_bucket = not args.bucket
 
             if is_pcluster_bucket:
+                # Validate the bucket prefix for both the default pcluster bucket
+                # and CustomS3Bucket if specified in the cluster configuration.
+                # Skip validation if a bucket is specified via the --bucket CLI argument.
                 self._validate_bucket_prefix(args.bucket_prefix)
 
             return self._export_cluster_logs(args, args.output_file)
@@ -89,7 +92,7 @@ class ExportClusterLogsCommand(ExportLogsCommand, CliCommand):
                 or bucket_prefix == PCLUSTER_BUCKET_PROTECTED_FOLDER
             ):
                 raise ValueError(
-                    f"Cannot export logs to {bucket_prefix} as it is a protected folder "
+                    f"Cannot export logs to {bucket_prefix} as it is within a protected folder "
                     f"in {PCLUSTER_BUCKET_PROTECTED_PREFIX}. Please use another folder."
                 )